requirements.txt
openai
python-dotenv
psycopg[binary]
grpcio-tools
pynng
requests
setuptools
